Extract additional util methods
